Get rid of LiveActionsController   * Fix: Remove liveactions.py - contained class LiveActionsController   * Add execute_action method to worker.py
